feat: avoid confussion with the upcoming `/term` command
Remove the `/privacy` and `/terms` commands in favor of the `/legal`
command to avoid confussion with the upcoming `/term` command.
